Moves axes to its own dedicated module
