+++ src/plotting.rs
@@ -1,3 +1,4 @@
+pub mod axes;
 pub mod canvas;
 pub mod colors;
 pub mod common;
